got basic json parser working
Doesn't really seem worth it to continue this since I would just be
better served using something like `ziggy` to read from a file, parse it
and cast to a known zig type.

Perhaps this is worth it if I provide a mechanism for the user to
provide a type of config to the "getConfig" part. Although this means
that I would basically just be wrapping `std.json` or some other
encoding.

....might abandon